% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_formula.R
\name{address_parser}
\alias{address_parser}
\title{Parse excel address including sheet}
\usage{
address_parser(address)
}
\arguments{
\item{address}{Character string of excel address, including sheet.}
}
\value{
list, with \verb{$sheet} giving sheet name, \verb{$rows} givng the row number, \verb{$cols} giving the column number.
}
\description{
Helper function to parse full excel address into sheet, row, column. Removes
\verb{'} from sheet names. If given a cell range, returns only the address of the first cell.
}
\examples{
address_parser("SPS!AS3")
address_parser("'Input Page'!$B$30")

}
