% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{as_numeric_safe}
\alias{as_numeric_safe}
\title{Safely convert character numerics into numeric}
\usage{
as_numeric_safe(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Either numeric vector of x or (if that is lossy) character returns input.
}
\description{
Function to translate columns that "should" be numeric into numeric
but doesn't do so if that would mean converting a true character string to NA.
}
