% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_chinook_input_template.R
\name{chinook_input_template_updater}
\alias{chinook_input_template_updater}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Update Chinook input template with values from the final TAMM of the previous year}
\usage{
chinook_input_template_updater(
  last_years_final_run,
  old_template,
  new_template_name,
  debug_mode = FALSE
)
}
\arguments{
\item{last_years_final_run}{filepath to last year's final run}

\item{old_template}{filepath to old template file}

\item{new_template_name}{filepath to save new template file}

\item{debug_mode}{Should updated cells be highlighted, and update regions be printed to console? Logical, defaults to FALSE.}
}
\value{
nothing
}
\description{
Takes (a) file path for the final run from the previous year (to take values FROM),
(b) the file path for the old template (to take formatting and non-changing cells FROM),
and (c) the filepath for the new template
}
\details{
Optionally, \code{debug_mode} can be set to TRUE. This will (a) print out the sheet and "to" addresses of each
transfer as they happen, and (b) highlight each cell in the template whose values are copied from \code{last_year_final_run}
in blood red, making it easy to check that the correct sections are being updated.
}
\examples{
\dontrun{
chinook_input_template_updater(last_years_final_run = here("Chin2225_final.xlsx"),
       old_template = here("Chinook_TAMM_Fishery_Input_Template_20241.xlsx"),
       new_template_name = here("Chinook_TAMM_Fishery_Input_Template_2025- CHECK.xlsx"),
                               debug_mode = TRUE)
chinook_input_template_updater(last_years_final_run = here("Chin2225_final.xlsx"),
       old_template = here("Chinook_TAMM_Fishery_Input_Template_20241.xlsx"),
       new_template_name = here("Chinook_TAMM_Fishery_Input_Template_2025.xlsx"))

}
}
