% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_formaters.R
\name{chunk_formater_percenter}
\alias{chunk_formater_percenter}
\title{Format chunks of dataframe to present as \%s, round digits}
\usage{
chunk_formater_percenter(df, block.ranges, percent.digits = 1)
}
\arguments{
\item{df}{dataframe of sheet to apply formatting to.}

\item{block.ranges}{vector of characters specifying blocks of cells (in excel nomenclature) to format as \%s}

\item{percent.digits}{Decimal place to round to in percent}
}
\value{
Formatted version of \code{df}
}
\description{
Format chunks of dataframe to present as \%s, round digits
}
