% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{chunk_read_2A_CUandM}
\alias{chunk_read_2A_CUandM}
\title{Read chunk for table 2A_CUandM}
\usage{
chunk_read_2A_CUandM(tamm_filepath, start_col, end_col, table_name)
}
\arguments{
\item{tamm_filepath}{tame file path}

\item{start_col}{Letter of the column the chunk starts on}

\item{end_col}{Letter of the column the chunk ends on}

\item{table_name}{name of table}
}
\description{
Read chunk for table 2A_CUandM
}
