% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{chunk_read_2A_CUandM_N}
\alias{chunk_read_2A_CUandM_N}
\title{Read in individual chunks for sheet 2A_Cu&M_N}
\usage{
chunk_read_2A_CUandM_N(full_sheet, start_col, end_col, table_name)
}
\arguments{
\item{full_sheet}{entire raw spreadsheet}

\item{start_col}{Letter of the column the chunk starts on}

\item{end_col}{Letter of the column the chunk ends on}

\item{table_name}{name of table}
}
\description{
Read in individual chunks for sheet 2A_Cu&M_N
}
