% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{chunk_read_2A_CUnmrkd}
\alias{chunk_read_2A_CUnmrkd}
\title{Reads chunk of 2A_CUnmrkd sheet}
\usage{
chunk_read_2A_CUnmrkd(tamm_filepath, start_col, end_col, table_name)
}
\arguments{
\item{tamm_filepath}{tame file path}

\item{start_col}{Letter of the column the chunk starts on}

\item{end_col}{Letter of the column the chunk ends on}

\item{table_name}{name of table}
}
\description{
Reads chunk of 2A_CUnmrkd sheet
}
