% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{chunk_read_finisher}
\alias{chunk_read_finisher}
\title{Finishes a chunk after initial reading}
\usage{
chunk_read_finisher(
  sheet_name,
  table_name,
  raw,
  raw_titles,
  data_er,
  do_er = TRUE
)
}
\arguments{
\item{sheet_name}{Name of sheet}

\item{table_name}{Name of table}

\item{raw}{.}

\item{raw_titles}{.}

\item{data_er}{.}

\item{do_er}{Should ER be done? If FALSE, skips ER (useful for processing sections with no ER.)}
}
\value{
A list with $aeq and $er, each a dataframe
}
\description{
Handles some of the repetitive tasks of processing individual chunks; this part is generalized
and does not need to be different for each sheet. Lots of fiddliness in here
because TAMM formatting is terrible
}
