% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_limiting_stock.R
\name{clean_limiting_stock}
\alias{clean_limiting_stock}
\title{Generates clean read of TAMM limiting stock sheet}
\usage{
clean_limiting_stock(filename)
}
\arguments{
\item{filename}{Name (and path) for TAMM files}
}
\value{
dataframe
}
\description{
Effectively a wrapper function for read_limiting_stock with some formatting
added in. Filters to unmarked naturals, just present ER values.
}
