% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_color_matches.R
\name{find_color_matches}
\alias{find_color_matches}
\title{Identify all cells in Excel sheet or sheets with matching cell colors}
\usage{
find_color_matches(file, target_cell, sheets = "Input Page")
}
\arguments{
\item{file}{Character string of excel file name, including path if relevant}

\item{target_cell}{Character string of excel address of cell with the cell shading of interest (e.g., "B4")}

\item{sheets}{character string or character vector of sheets to search through}
}
\value{
tibble summarizing entries with the color of interest, including sheet, row and column ids, and row name (entry of first column in sheet)
}
\description{
Designed for tracking the status of TAMM inputs, which are typically color coded.
}
\examples{
\dontrun{
file = "chin tamms/Chin2023_Final.xlsx"
find_color_matches(file, "B20")
}
}
