% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishery_chinook_fram}
\alias{fishery_chinook_fram}
\title{Chinook fishery information}
\format{
A data frame with 74 rows and 5 columns:
\describe{
\item{species}{Species name}
\item{version_number}{}
\item{fishery_id}{fishery id number in FRAM}
\item{fishery_name}{fishery name in FRAM}
\item{fishery_title}{consistent and more human readable version of \code{fishery_name}}
}
}
\source{
2024 Pre-Season Chinook DB.mdb
}
\usage{
fishery_chinook_fram
}
\description{
Mapping of fishery_id to fishery name for Chinook salmon, taken from the \code{Fishery} table of
the Chinook FRAM database.
}
\keyword{datasets}
