% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishery_coho_fram}
\alias{fishery_coho_fram}
\title{Coho fishery information}
\format{
A data frame with 198 rows and 5 columns:
\describe{
\item{species}{Species name}
\item{version_number}{}
\item{fishery_id}{fishery id number in FRAM}
\item{fishery_name}{fishery name in FRAM}
\item{fishery_title}{consistent and more human readable version of \code{fishery_name}}
}
}
\source{
2024NOF_CohoFRAMdatabase_distribution.mdb
}
\usage{
fishery_coho_fram
}
\description{
Mapping of fishery_id to fishery name for Coho salmon, taken from the \code{Fishery} table of
the Coho FRAM database.
}
\keyword{datasets}
