% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_renamer.R
\name{fishery_renamer}
\alias{fishery_renamer}
\title{relabel inconsistent fishery names in FRAM/TAMM}
\usage{
fishery_renamer(x, sep = FALSE)
}
\arguments{
\item{x}{character vector of fishery names from TAMM or FRAM}

\item{sep}{Should we return a dataframe with separate column for class of fishing? Defaults to FALSE}
}
\value{
#vector with cleaned fishery names OR dataframe with cleaned name (\verb{$full}),
area without fishery type (\verb{$area}) and type of fishery (\verb{$class}).
}
\description{
With argument \code{sep = TRUE}, will return data frame with separate
columns for the area and the the "class" (Net, Troll, Sport)
}
