% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_key_tamm_sheets.R
\name{format_key_tamm_sheets}
\alias{format_key_tamm_sheets}
\title{Modify list of TAMM spreadsheet dataframes to facilitate comparison.}
\usage{
format_key_tamm_sheets(dat, percent.digits = 1, numeric.digits = 1)
}
\arguments{
\item{dat}{list of dataframes corresponding to the overview, limiting stock, and inputs tabs. Must be named \verb{$overview}, \verb{$limiting}, and \verb{$input}}

\item{percent.digits}{Optional, number of decimal places to round percentages to before comparison. Defaults to 1.}

\item{numeric.digits}{Optional, number of decimal places to round non-percentage numerics to before comparison. Defaults to 1.}
}
\value{
list of dataframes with same structure as \code{dat}, contents modified.
}
\description{
Modify list of TAMM spreadsheet dataframes to facilitate comparison.
}
