% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_overview.R
\name{format_overview}
\alias{format_overview}
\title{Format TAMM overview for easy printing}
\usage{
format_overview(dat.overview)
}
\arguments{
\item{dat.overview}{dataframe of TAMM overview page, as produced by read_overview()}
}
\value{
list with the formatted data, a vector of indent information, and a vector of bolding information.
}
\description{
Format TAMM overview for easy printing
}
