% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_formula.R
\name{formula_formater}
\alias{formula_formater}
\title{Add spaces to excel formulas for readability}
\usage{
formula_formater(x, newline_breakpoint = NULL)
}
\arguments{
\item{x}{Character string containing an excel formula}

\item{newline_breakpoint}{Pattern to identify locations for newlines. Could be a single character (e.g. \code{"/"})
or a more complex regular expression (e.g., \code{"[)][*-+/]"} to add a newline after a combination of an end parathesis and an operator).}
}
\value{
character string
}
\description{
Helper function to add spaces to excel formulas to improve readability. Can also
add newlines based on a character or regular expression
}
\examples{
formula_formater("(SUM(A1:A4)*5/11)/(B3-B4)")
formula_formater("(SUM(A1:A4)*5/11)/(B3-B4)", newline_breakpoint = "/")
formula_formater("(SUM(A1:A4)*5/11)/(B3-B4)", newline_breakpoint = "[*-+/]")

}
