% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_formaters.R
\name{fun_percenter}
\alias{fun_percenter}
\title{Format vector of mixed characters to present numbers as percents}
\usage{
fun_percenter(x, percent.digits)
}
\arguments{
\item{x}{Character vector, presumably containing some entries that are numbers in character form}

\item{percent.digits}{Number of digits to round to after converting to percents}
}
\value{
character vector with individual entries converted to percentages and rounded as appropriate.
}
\description{
Intended for internal use within formatting functions. Note that percents in excel are read in
as proportions in R -- this makes them percents to make saved files more readable.
}
