% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_formaters.R
\name{fun_rounder}
\alias{fun_rounder}
\title{Format vector of mixed characters to round numbers to specified digits}
\usage{
fun_rounder(x, digits)
}
\arguments{
\item{x}{Character vector, presumably containing some entries that are numbers in character form}

\item{digits}{Number of digits to round numeric items to}
}
\description{
Intended for internal use within formatting functions.
}
