% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_overview.R
\name{kable_overview}
\alias{kable_overview}
\title{Print table of overview using kable and kableExtra}
\usage{
kable_overview(dat_overview, ind_ident, col_bold)
}
\arguments{
\item{dat_overview}{formatted data, first item of format_overview() output}

\item{col_bold}{vector of names for bolding to match TAMM overview formatting. Third item of format_overview() output.}

\item{ind_indent}{Vector of indices that need indenting to match TAMM overview formatting. Second item of format_overview() output}
}
\value{
html table
}
\description{
Uses output of format_overview()
}
