% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_formula.R
\name{make_tracer_edges}
\alias{make_tracer_edges}
\title{Calculate and format edges from traced objects}
\usage{
make_tracer_edges(address_list)
}
\arguments{
\item{address_list}{The \verb{$raw.tracing} component of the results of a \code{trace_formula()} call.}
}
\value{
tibble.
\verb{$from} is the index of the cell that is referenced,
\verb{$to} is hte index of the cell that is doing the referencing
\verb{$from_name} and \verb{$to_name} are the excel addresses of the same.
}
\description{
Helper function to find all references in list created within tracer_formula, and create a simple dataframe
representing all the "edges" of the traced network (that is, all the connections between cells).
Gets called in \code{trace_formula()} to generate the \verb{$references} component of the output.
}
