% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_formula.R
\name{make_tracer_network}
\alias{make_tracer_network}
\title{Visualize excel formula tracing}
\usage{
make_tracer_network(tracer_list, save_path = NULL, newline_breakpoint = NULL)
}
\arguments{
\item{tracer_list}{Output of a \code{trace_formula()} call.}

\item{save_path}{Optional, defaults to NULL. Provide a file path + name to save an html of the visualization. Make sure file name ends in \code{.html}.}

\item{newline_breakpoint}{Optional character string used to add breakpoints to formulas to improve readability. Can be simple character or regular expression; See \code{formula_formater} for details and examples.}
}
\value{
interactive visNetwork object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Render the dependency network created in \code{trace_formula()}; optionally save this to an interactive .html.
}
\examples{
\dontrun{
trace_network = trace_formula(path = "NOF material/NOF 2024/NOF 2/Chin1624.xlsx",
cell_start = "SPSmrkd!AS20")
make_tracer_network(trace_network)
}
}
