% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_formula.R
\name{make_tracer_nodes}
\alias{make_tracer_nodes}
\title{List cells and contents from traced objects}
\usage{
make_tracer_nodes(address_list)
}
\arguments{
\item{address_list}{The \verb{$raw.tracing} component of the results of a \code{trace_formula()} call.}
}
\value{
tibble summarizing each cell in the tracing. See \code{trace_formula()} for details.
}
\description{
Helper function to parse list created within trace_formula into a tibble with each cell represented by a row.
Called in \code{trace_formula()} to generate the \verb{$cells} component of the output.
}
