% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{process_2ac_aeqs}
\alias{process_2ac_aeqs}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Read and combine aeq sections form tables 2a-2c}
\usage{
process_2ac_aeqs(tab_2a, tab_2b, tab_2c)
}
\arguments{
\item{tab_2a}{the "Table 2a" section of one of the 2ac sheets, read in with
a \verb{chunk_reader_2A_*()} function}

\item{tab_2b}{same, but 2b}

\item{tab_2c}{same, but 2c}
}
\value{
dataframe of AEQ values for each stock in longform
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Read and combine aeq sections form tables 2a-2c
}
