% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_overview.R
\name{quick_er_format}
\alias{quick_er_format}
\title{Quick helper function to format vectors that contain text and ERs}
\usage{
quick_er_format(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Quick helper function to format vectors that contain text and ERs
}
