% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{read_2a_sheets}
\alias{read_2a_sheets}
\title{Read and combine all 2a sheets from a TAMM file}
\usage{
read_2a_sheets(tamm_filepath, stock_cleanup = TRUE)
}
\arguments{
\item{tamm_filepath}{TAMM filepath}

\item{stock_cleanup}{Clean up stock name capitalization? Logical, defaults to TRUE.}
}
\value{
list, with $aeq containing the primary information (AEQs of fishery impacts on each stock) and $er containing the ER information at the bottom of the spreadsheets.
}
\description{
Gathers and cleans the AEQ and ER information for tamm sheets \verb{2aCmrkd},
\verb{2aCUnmkrd}, \verb{2aCUandM}, and \verb{2aCUandM_N}.
}
\examples{
\dontrun{
library(here)
filepath <- here("tamms/Chin2025.xlsx")
res = read_2a_sheets(filepath)
jdf = read_jdf(filepath)
}
}
