% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{read_2aCmrkd}
\alias{read_2aCmrkd}
\title{Reads sheet of 2A_Cmrkd}
\usage{
read_2aCmrkd(tamm_filepath, stock_cleanup = TRUE)
}
\arguments{
\item{tamm_filepath}{path to file}

\item{stock_cleanup}{Do cleanup of capitalization?}
}
\value{
a list, with $aeq (the main parts of each table) and $er (the very bottom section)
}
\description{
Reads sheet of 2A_Cmrkd
}
