% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{read_2aCUandM}
\alias{read_2aCUandM}
\title{Read and process all parts of sheet 2A CU and M}
\usage{
read_2aCUandM(tamm_filepath, stock_cleanup = TRUE)
}
\arguments{
\item{tamm_filepath}{Tamm filepath}

\item{stock_cleanup}{Clean up weird capitalization? Logical, defaults to TRUE}
}
\value{
a list, with $aeq (the main parts of the table) and $er (the very bottom section) @return
}
\description{
Read and process all parts of sheet 2A CU and M
}
