% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{read_2aCUandM_N}
\alias{read_2aCUandM_N}
\title{Read and combine all chunks for sheet 2A CU and M_N}
\usage{
read_2aCUandM_N(tamm_filepath, stock_cleanup = TRUE)
}
\arguments{
\item{tamm_filepath}{Tamm filepath}

\item{stock_cleanup}{Clean up stock name capitalization? Logical, defaults to TRUE}
}
\description{
Read and combine all chunks for sheet 2A CU and M_N
}
