% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{read_2aCUnmrkd}
\alias{read_2aCUnmrkd}
\title{read all info for 2aCUnmrkd and process it.}
\usage{
read_2aCUnmrkd(tamm_filepath, stock_cleanup = TRUE)
}
\arguments{
\item{tamm_filepath}{Tamm filepath}

\item{stock_cleanup}{Logical, defaults to TRUE. Convert weird capitalization to lowercase?}
}
\value{
a list, with $aeq (the main parts of the table) and $er (the very bottom section)
}
\description{
read all info for 2aCUnmrkd and process it.
}
