% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc_sheets.R
\name{read_jdf}
\alias{read_jdf}
\title{Read in AEQ and ER information from the JDF sheet}
\usage{
read_jdf(tamm_filepath, stock_cleanup = TRUE)
}
\arguments{
\item{tamm_filepath}{TAMM filepath}

\item{stock_cleanup}{Clean up stock name capitalization? Logical, defaults to TRUE.}
}
\value{
List, with $aeq containing the combined AEQ information that is spread across multiple chunks in the TAMM sheet. List also contains $er = NULL, to provide parallel outputs to \code{read_2a_sheets}, which \emph{does} have a non-null $er item in the output list.
}
\description{
Read in AEQ and ER information from the JDF sheet
}
\examples{
\dontrun{
library(here)
filepath <- here("tamms/Chin2025.xlsx")
jdf = read_jdf(filepath)
}
}
