% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_key_tamm_sheets.R
\name{read_key_tamm_sheets}
\alias{read_key_tamm_sheets}
\title{Read TAMM files to extract the key sheets}
\usage{
read_key_tamm_sheets(xlsxFile)
}
\arguments{
\item{xlsxFile}{#Tamm file name}
}
\value{
List of dataframes: \verb{$overview}, \verb{$limiting}, and \verb{$input}
}
\description{
Read TAMM files to extract the key sheets
}
