% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_key_tamm_sheets.R
\name{read_key_tamm_sheets_chin}
\alias{read_key_tamm_sheets_chin}
\title{Read Chinook TAMM files to extract the key sheets}
\usage{
read_key_tamm_sheets_chin(file)
}
\arguments{
\item{file}{#Tamm file name}
}
\value{
List of dataframes: \verb{$overview}, \verb{$limiting}, and \verb{$input}
}
\description{
Read Chinook TAMM files to extract the key sheets
}
