% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_key_tamm_sheets.R
\name{read_key_tamm_sheets_coho}
\alias{read_key_tamm_sheets_coho}
\title{Read Chinook TAMM files to extract the key sheets}
\usage{
read_key_tamm_sheets_coho(xlsxFile)
}
\arguments{
\item{xlsxFile}{#Tamm file name}
}
\value{
List of dataframes: \verb{$two}, \verb{$tami}, and \verb{$wacoast}
}
\description{
Read Chinook TAMM files to extract the key sheets
}
