% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_limiting_stock.R
\name{read_limiting_stock}
\alias{read_limiting_stock}
\title{Read TAMM limiting stock complete tab}
\usage{
read_limiting_stock(filename, longform = FALSE)
}
\arguments{
\item{filename}{Tamm name (and path)}

\item{longform}{Should results be in long form (good for R stuff) (\code{TRUE}) or replicate the structure of the TAMM sheet (\code{FALSE}). Logical, defaults to \code{FALSE}.}
}
\value{
data frame summarizing the TAMM limiting stock tab.
}
\description{
Reads in the table, adds fishery type information.
}
