% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_management.R
\name{read_management_chin}
\alias{read_management_chin}
\title{Extract and format management criterion from Chinook TAMM}
\usage{
read_management_chin(file)
}
\arguments{
\item{file}{filename, including filepath if appropriate}
}
\value{
List of stock lists
}
\description{
Extract and format management criterion from Chinook TAMM
}
\examples{
\dontrun{
cur.management = read_management_chin("Code Inputs/Pre-Season/TAMMs/2013_Final W_BP7.1.xlsx")
}
}
