% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_overview.R
\name{read_overview_complete}
\alias{read_overview_complete}
\title{Read overview sheet from TAMM and return all stock}
\usage{
read_overview_complete(path)
}
\arguments{
\item{path}{filename (including path) to Chinook TAMM}
}
\value{
Dataframe containing all infomration from cells A2:H34 of the "ER_ESC_Overview_New" sheet.
}
\description{
The stock column of the overview tab uses R-unfriendly approaches to store
season, overall stock name, and individual stocks in the \code{stock} column of the TAMM. \code{read_overview_complete}
splits these into separate columns: \code{season}, \code{primary_stock}, and \code{stock}.
For three stock (Green, Puyallup, Skokomish), the tamm presents model predictions across two rows with
a single stock name due to merged cells. The \code{stock} column in the output preserves this information;
the entry corresponding to the second row has the suffix "_row2".
}
