% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tnt_allocation_chin.R
\name{read_tnt_allocation_chin}
\alias{read_tnt_allocation_chin}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
read_tnt_allocation_chin(xlsxFile)
}
\arguments{
\item{xlsxFile}{Character vector. Filename (including path) for chinook TAMM}
}
\value{
dataframe with the treaty and nontreaty mortalities at each stock.
}
\description{
Extract Treaty and Nontreaty numbers from "2A_CU&M_H+N" TAMM sheet. For Elwha and Dungeoness,
separate allocations are calculated using column O of the JDF tab
}
\details{
Currently provides dataframe with \code{stock.original} column identifying the stock based on their names
in the TAMM, and \code{stock.clean} with more human-readable names (which are generally consistent
with other stock names, like those used to label the management criterion).
}
