% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_management.R
\name{reformat_management}
\alias{reformat_management}
\title{Convenience function for formatting TAMM management criterion chunks}
\usage{
reformat_management(df, cols.percent, rows.header = 2, notes = "")
}
\arguments{
\item{df}{dataframe trimmed from TAMM management page to just the relevant stock's criterion.}

\item{cols.percent}{Column numbers for columns that should be converted to percents.}

\item{rows.header}{How many rows are used to make up the header? Defaults to 2, but sometimes header block
in TAMM is 1 row.}

\item{notes}{Optional. Character vector, each item of which is a separate note
associated with this stock. Generally based on comments in excel doc, can also feed in "extra" rows from TAMM that
contain caveats.}
}
\value{
Stock managament criterion list, containing dataframe extracted from TAMM and notes hard-coded in based on TAMM cell comments.
}
\description{
Internal tool to simplify converting TAMM sheet to stock management criterion lists. See
\code{read_management_chin()} for example use.
}
