% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_formula.R
\name{safe_convert_numeric}
\alias{safe_convert_numeric}
\title{Convert string of number to numeric}
\usage{
safe_convert_numeric(x)
}
\arguments{
\item{x}{character string}
}
\value{
Input formatted as character string or numeric
}
\description{
Helper function to check an atomic string and convert to numeric if
it contains a legal number.
}
\examples{
safe_convert_numeric("AS20-BC2")
safe_convert_numeric("10.5")
}
