% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_chinook_fram}
\alias{stock_chinook_fram}
\title{Chinook stock information}
\format{
A data frame with 78 rows and 7 columns:
\describe{
\item{species}{Species name}
\item{stock_version}{}
\item{stock_id}{Chinook stock id number in FRAM}
\item{production_region_number}{}
\item{management_unit_number}{}
\item{stock_name}{Chinook stock name in FRAM}
\item{stock_long_name}{\code{stock_name} but more human readable}
}
}
\source{
2024 Pre-Season Chinook DB.mdb
}
\usage{
stock_chinook_fram
}
\description{
Mapping of stock_id to stock name for Chinook salmon, taken from the \code{Stock} table of
the Chinook FRAM database.
}
\keyword{datasets}
