% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamm_compare.R
\name{tamm_compare}
\alias{tamm_compare}
\title{Generate summary comparison of any number of TAMM files}
\usage{
tamm_compare(
  tamm.names,
  tamm.path = getwd(),
  fisheries = NULL,
  clean = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{tamm.names}{character vector of the three tamm files to compare (including .xlsx suffix).}

\item{tamm.path}{Absolute path to directory containing the tamm file. \code{here::here()} can be useful in identifying appropriate path. Character atomic; defaults to current working directory.}

\item{fisheries}{Optional, numeric vector of fishery IDs to filter to before plotting fishery-specific ER for each stock. Defaults to NULL (no filtering).}

\item{clean}{Should the intermediate .qmd files used to make the report be deleted afterwards? Logical, defaults to \code{TRUE}. Set to \code{FALSE}
to explore the .qmd files underlying the report.}

\item{overwrite}{Should the intermediate .qmd files or the final report be overwritten if those files already exist?; Logical, defaults to \code{TRUE}.}
}
\value{
Nothing
}
\description{
Intended use case is to compare low, mid, and high ocean option.
}
\examples{
\dontrun{
tamm.path = "C:/Users/edwc1477/Documents/WDFW FRAM team work/NOF material/NOF 2024/FRAM"
tamm.names =c("Chin1024.xlsx", "Chin1124.xlsx", "Chin1224.xlsx")
tamm_compare3(tamm.names = tamm.names, tamm.path = tamm.path)
}
}
