% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamm_diff.R
\name{tamm_diff_coho}
\alias{tamm_diff_coho}
\title{Compare key sheets of Coho TAMM files}
\usage{
tamm_diff_coho(
  filename.1,
  filename.2,
  results.name,
  percent.digits = 1,
  numeric.digits = 1,
  numeric.digits.small = 4
)
}
\arguments{
\item{filename.1}{name of first TAMM file to compare. Include file path if file is not in working directory.}

\item{filename.2}{name of second TAMM file to compare. Include file path if file is not in working directory.}

\item{results.name}{name of output sheets. Include file path if save location is not in working directory.}

\item{percent.digits}{Number of decimals to round percentages to before comparing. Defaults to 1.}

\item{numeric.digits}{Number of decimals to round numbers to before comparing. Applied to cells which expect to be whole numbers (e.g. #s of fish). Defaults to 1.}

\item{numeric.digits.small}{Number of decimals to round numbers to before comparing. Applied to cells which expect to be small decimals. Defaults to 4.}
}
\description{
Compare key sheets of Coho TAMM files
}
