% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamm_report.R
\name{tamm_report}
\alias{tamm_report}
\title{Generate report of figures from TAMM file}
\usage{
tamm_report(tamm.path, tamm.name, clean = TRUE)
}
\arguments{
\item{tamm.path}{Absolute path to directory containing the tamm file. (hoping to streamline this at some point). Character atomic.}

\item{tamm.name}{Name of tamm file (including .xlsx suffix). Character atomic}

\item{clean}{Should the .qmd files used to make the report be deleted afterwards? Logical, defaults to \code{TRUE}. Set to \code{FALSE}
to explore the .qmd files underlying the report.}
}
\value{
Nothing
}
\description{
Generate report of figures from TAMM file
}
\examples{
\dontrun{
tamm.path = "C:/Users/edwc1477/Documents/WDFW FRAM team work/NOF material/NOF 2024/NOF 2"
tamm.name = "Chin2124.xslx"
tamm_report(tamm.path = tamm.path, tamm.name = "Chin2124.xlsx")
}
}
