% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAMMsupport-package.R
\docType{package}
\name{TAMMsupport-package}
\alias{TAMMsupport}
\alias{TAMMsupport-package}
\title{TAMMsupport: Streamline working with Terminal Area Management Modules}
\description{
A convenient tool for interfacing with Terminal Area Manamagement Modules (TAMMs) in R environments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://framverse.r-universe.dev/TAMMsupport}
  \item \url{https://github.com/cbedwards-dfw/TAMMsupport}
  \item \url{https://cbedwards-dfw.github.io/TAMMsupport/}
}

}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

Authors:
\itemize{
  \item Ty Garber \email{tyler.garber@dfw.wa.gov}
}

}
\keyword{internal}
