% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timestep_chinook_fram}
\alias{timestep_chinook_fram}
\title{Chinook timestep information stock information}
\format{
A data frame with 4 rows and 5 columns:
\describe{
\item{species}{Species name}
\item{version_number}{}
\item{time_step_id}{id number for the time step}
\item{time_step_name}{Span of each timestep. Timesteps start on the first of the month, and end on the last of the month. Note that \code{Oct-Apr-2} is highlighting that timestep 4 runs from October of the current year to April of the NEXT year (equivalent to timestep 1 of the following year).}
\item{time_step_title}{\code{time_step_name}, but months are written out}
}
}
\source{
2024 Pre-Season Chinook DB.mdb
}
\usage{
timestep_chinook_fram
}
\description{
Mapping of timestep numbers to dates in the year, taken from the \code{TimeStep} table of
the Chinook FRAM database
}
\keyword{datasets}
