% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timestep_coho_fram}
\alias{timestep_coho_fram}
\title{Coho timestep information stock information}
\format{
A data frame with 5 rows and 5 columns:
\describe{
\item{species}{Species name}
\item{version_number}{}
\item{time_step_id}{id number for the time step}
\item{time_step_name}{Span of each timestep. Timesteps start on the first of the month, and end on the last of the month.}
\item{time_step_title}{\code{time_step_name}, but months are written out consistently}
}
}
\source{
2024NOF_CohoFRAMdatabase_distribution.mdb
}
\usage{
timestep_coho_fram
}
\description{
Mapping of timestep numbers to dates in the year, taken from the \code{TimeStep} table of
the Chinook FRAM database
}
\keyword{datasets}
