% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_data_frame}
\alias{validate_data_frame}
\title{Confirm object is dataframe}
\usage{
validate_data_frame(
  x,
  ...,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{Object to check}

\item{...}{additional arguments}

\item{arg}{name of argument in calling function, identified programmatically using rlang}

\item{call}{name of calling function, identified programmatically using rlang}
}
\description{
Confirm object is dataframe
}
