% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_chinook_inputs.R
\name{compare_chinook_inputs}
\alias{compare_chinook_inputs}
\title{Compare differences in input sheets of two TAMM files}
\usage{
compare_chinook_inputs(
  file1,
  file2,
  digits.signif = 4,
  trim.cols = TRUE,
  diff.only = TRUE
)
}
\arguments{
\item{file1}{Filename of first TAMM file}

\item{file2}{Filename of second TAMM file}

\item{digits.signif}{How many digits should be rounded to before looking for differences? Without this, miniscule excel numerical differences (e.g. differing by 0.000000001) can get picked up and overwhelm the meaningful differences.}

\item{trim.cols}{Should columns that contain no differences be cut from the results? Defaults to TRUE.}

\item{diff.only}{Should entries that do not contain differences be replaced by "" to help spot differences? Defaults to TRUE.}
}
\value{
Tibble of differences in the input files; only rows with differences are included. \code{row_name} variable gives the row name in the TAMM file, and column names match excel column names.
}
\description{
Summarizes the differences in input sheets and returns them as a tibble.
}
