% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_limiting_stock.R
\name{read_limiting_stock}
\alias{read_limiting_stock}
\title{Read TAMM limiting stock complete tab}
\usage{
read_limiting_stock(file, longform = FALSE)
}
\arguments{
\item{file}{Tamm name (and path)}

\item{longform}{Should results be in long form (good for R stuff) (\code{TRUE}) or replicate the structure of the TAMM sheet (\code{FALSE}). Logical, defaults to \code{FALSE}.}
}
\value{
data frame summarizing the TAMM limiting stock tab. The "block" of the limiting tab is identified with column "stock_type". "first_section" is the first section in the TAMM; this section is primarily natural fish but includes some hatchery fish if they were included in management objectives. "ALL_H" is all hatchery fish (in TAMM, block starts on row 81), "UM_H" is unmarked hatchery (in TAMM, starts on row 160), "UM_N" is unmarked naturals (in TAMM, starts row 239), "AD_H" is marked hatchery (in TAMM, starts row 318), and "AD_N" is marked naturals (unless programs begin marking natural fish, this should always be all 0s; in TAMM starts row 397).
}
\description{
Reads in the table, adds fishery type information.
}
