% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_limiting_stock.R
\name{read_limiting_stock}
\alias{read_limiting_stock}
\title{Read TAMM limiting stock complete tab}
\usage{
read_limiting_stock(filename)
}
\arguments{
\item{filename}{Tamm name (and path)}
}
\value{
data frame summarizing the TAMM limiting stock tab.
}
\description{
Reads in the table, adds fishery type information.
}
