% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tnt_allocation_chin.R
\name{read_tnt_allocation_chin}
\alias{read_tnt_allocation_chin}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
read_tnt_allocation_chin(xlsxFile)
}
\arguments{
\item{xlsxFile}{Character vector. Filename (including path) for chinook TAMM}
}
\value{
dataframe with the treaty and nontreaty mortalities at each fishery.
}
\description{
Extract Treaty and Nontreaty numbers from "2A_CU&M_H+N" TAMM sheet. For Elwha and Dungeoness,
separate allocations are calculated using column O of the JDF tab
}
\details{
Currently provides dataframe with \code{fishery.original} column identifying the fishery.
With advice from Derek, will add a \code{fishery} column that is human-readable version of above. Might
Try to create same naming conventions as in \code{read_overview_complete()}.
}
