% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamm_diff.R
\name{tamm_diff}
\alias{tamm_diff}
\title{Compare key sheets of Chinook TAMM files}
\usage{
tamm_diff(filename.1, filename.2, results.name)
}
\arguments{
\item{filename.1}{name of first TAMM file to compare. Include file path if file is not in working directory.}

\item{filename.2}{name of second TAMM file to compare. Include file path if file is not in working directory.}

\item{results.name}{name of output sheets. Include file path if save location is not in working directory.}
}
\description{
Compare key sheets of Chinook TAMM files
}
\examples{
\dontrun{
tamm_diff(filename.1 = here("FRAM/Chin1124.xlsx"),
filename.2 = here("NOF 2/Chin2524.xlsx"),
results.name = here("Chin 1124 vs Chin 2524.xlsx")
)
}
}
