% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamm_report.R
\name{tamm_report}
\alias{tamm_report}
\title{Generate report of figures from TAMM file}
\usage{
tamm_report(
  tamm_name,
  tamm_path = getwd(),
  clean = TRUE,
  overwrite = TRUE,
  additional_children = NULL,
  additional_support_files = NULL
)
}
\arguments{
\item{tamm_name}{Name of tamm file (including .xlsx suffix). Character atomic}

\item{tamm_path}{Absolute path to directory containing the tamm file. \code{here::here()} can be useful in identifying appropriate path. Character atomic; defaults to current working directory.}

\item{clean}{Should the intermediate .qmd files used to make the report be deleted afterwards? Logical, defaults to \code{TRUE}. Set to \code{FALSE}
to explore the .qmd files underlying the report.}

\item{overwrite}{Should the intermediate .qmd files or the final report be overwritten if those files already exist?; Logical, defaults to \code{TRUE}.}

\item{additional_children}{Optional argument with filepath(s) to additional quarto child documents (see \code{Details}). Defaults to \code{NULL}.}

\item{additional_support_files}{Optional argument with filepath(s) to additional files needed by additional quarto documents. Files will be copied into the same directory as the TAMM for
easy reading/use by \code{additional_children} quarto files, and then deleted after the report is generated}
}
\value{
Nothing
}
\description{
Generate report of figures from TAMM file
}
\details{
This function generates a summary report of a single chinook TAMM (\href{https://framverse.github.io/fram_doc/calcs_data_chin.html#5_Terminal_Area_Management_Module}{Terminal Area Management Module}) file.
The common use case is to call this function for a tamm, and then view the resulting html report (which will be generated in the same folder as the TAMM).
However, this function also provides substantial flexibility in the form of optional user-created child quarto documents, which requires
some additional context to develop and use.

\code{tamm_report} is implemented using a parameterized quarto report and a child quarto document, which are included in the package.
The resulting report includes some broadly useful content -- currently, a replication of the overview TAMM sheet and
bar charts showing the breakdown of exploitation rates by fishery for each stock listed in the
\verb{limiting stock complete} TAMM sheet. However, individuals or organizations may consistently want
additional, specific visualizations based on their own needs. For example, individual tribes may want
to visualize impacts on a single stock, but use AEQ instead of ER rates. The \code{tamm_report} function is
designed to integrate these extra components with the \code{additional_children} argument.

The basic principle is that the individual or group can write a reusable "child" .qmd file, \href{https://bookdown.org/yihui/rmarkdown-cookbook/child-document.html}{see this explanation using Rmarkdown}), which will effectively
be inserted into the main report. As an example, we might create a simple additional
child .qmd file called \code{extra-tamm-child.qmd} to include a small table summarizing a few
key aspects of the Nooksack Earlies stock. The contents of that file are provided in the example below.

When we call \code{tamm_report()}, we can include as an argument \verb{additional_children = } with the file name (including file path) for
our new file. In this way we can easily include this extra content any time we create a tamm report. If desired,
multiple child files can be created, and \code{additional_children} can take a character vector with each child file name.

When writing child documents, formatting and chunks work akin to copy-pasting a section of a quarto file out of the main file. (To easily
access the main quarto file, run tamm_report() with \code{clean = FALSE}, which will leave \code{tamm-visualizer.qmd} and \code{tamm-visualizer-child.qmd} files in the folder with the TAMM file).
The child document will have access to key objects
\itemize{
\item \code{dat.all}: the output of \code{\link{read_limiting_stock}} with \code{longform = TRUE} called on the TAMM;
\item \code{dat_overview}: the output of of \code{\link{read_overview}} called on the TAMM;
\item and \code{dat}: the specific filtered and formatted version of the limiting stock data generated in the \verb{limiting stock tab} chunk of \code{tamm-visualizer.qmd} and used in generating the unmarked natural stock exploitation rate figures.
When developing a child document, it may be useful to create in the working environment versions of \code{dat.all} and \code{dat_overview} from an example TAMM to facilitate
writing appropriate filters, plot-making code, etc. The primary quarto report loads in the tidyverse, framrsquared, kableExtra, and TAMMsupport packages,
but you can include additional packages for the child documents as needed by adding \verb{library(packagename} calls just
like in any other context.
}

If the \code{additional_children} files need to read contents of additional files (or call children .qmd files of their own
to streamline looping over stocks or fisheries), the file paths to these supplemental
files can be provided in the \verb{additional.support files} argument. These files will be copied into the
same folder as the .qmd files, so the children quarto files can be written to access the supplemental
files without dealing with file paths.

For questions/support in developing child documents to extend the report, contact Collin Edwards at WDFW.
}
\examples{
\dontrun{
##### CONTENTS FOR extra-tamm-child.qmd #####-----------------------------
## Test section

Hypothetical world in which we want a table summarizing the impacts of three
fisheries on the "Nooksack Earlies" stock.

```{r}
kable(dat |> filter(stock == "Nooksack Earlies") |> filter(fishery_id \%in\% 16:18))
```
#### END CONTENTS FOR extra-tamm-child.qmd #####--------------------------

tamm_path = "C:/Users/edwc1477/Documents/WDFW FRAM team work/NOF material/NOF 2024/NOF 2"
tamm_name = "Chin2124.xslx"
tamm_report(tamm_name = "Chin2124.xlsx", tamm_path = tamm_path)
## Now with our extra content, presuming `extra-tamm-child.qmd` is saved in our documents folder
tamm_report(tamm_name = "Chin2124.xlsx", tamm_path = tamm_path,
  additional_children = "C:/Users/JohnDoe/Documents/extra-tamm-child.qmd")
}
}
