% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamm_report.R
\name{tamm_report}
\alias{tamm_report}
\title{Generate report of figures from TAMM file}
\usage{
tamm_report(tamm.name, tamm.path = getwd(), clean = TRUE, overwrite = TRUE)
}
\arguments{
\item{tamm.name}{Name of tamm file (including .xlsx suffix). Character atomic}

\item{tamm.path}{Absolute path to directory containing the tamm file. \code{here::here()} can be useful in identifying appropriate path. Character atomic; defaults to current working directory.}

\item{clean}{Should the intermediate .qmd files used to make the report be deleted afterwards? Logical, defaults to \code{TRUE}. Set to \code{FALSE}
to explore the .qmd files underlying the report.}

\item{overwrite}{Should the intermediate .qmd files or the final report be overwritten if those files already exist?; Logical, defaults to \code{TRUE}.}
}
\value{
Nothing
}
\description{
Generate report of figures from TAMM file
}
\examples{
\dontrun{
tamm.path = "C:/Users/edwc1477/Documents/WDFW FRAM team work/NOF material/NOF 2024/NOF 2"
tamm.name = "Chin2124.xslx"
tamm_report(tamm.name = "Chin2124.xlsx", tamm.path = tamm.path)
}
}
